dont' replace all matches!

+++ src/transformers/models/llama/configuration_llama.py
@@ -61,7 +61,7 @@ class LlamaConfig(PreTrainedConfig):
         "norm": (["hidden_states"], ["hidden_states"]),
     }
 
-    vocab_size: float = 32000
+    vocab_size: int = 32000
     hidden_size: int = 4096
     intermediate_size: int = 11008
     num_hidden_layers: int = 32
